+++ app/agent/page.tsx
@@ -1,0 +1,5 @@
+import { redirect } from "next/navigation";
+
+export default function Agent() {
+  redirect(`./agent/${crypto.randomUUID()}`);
+}
