--- app/chat/page.tsx
@@ -1,5 +1,0 @@
-import { redirect } from "next/navigation";
-
-export default function Chat() {
-  redirect(`./chat/${crypto.randomUUID()}`);
-}
